pyproject.toml
[tool.poetry]
name = "drf by zifirпnka"
version = "0.1.0"
description = "Проект на Django REST Framework"
authors = ["Student <ananimuszahar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"
nullable = "^0.2.0"
stripe = "^11.6.0"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
